nextButton to child component
