+++ src/components/answerOptions/answerOptions.module.scss
@@ -78,39 +78,4 @@
     }
 }
 
-.Disabled {
-    display:inline-block;
-    margin-bottom: 2rem;
-    color: $white;    
-    font-family: "Lato", Arial, sans-serif;
-    font-size: 0.9375rem;
-    font-weight: 400;
-    border-radius: 0.25rem;
-    border: 1px solid $dark-mode-border;    
-    background-color: $dark;
-    box-shadow: none;
-    width: 100%;
-    cursor: not-allowed;
-    line-height: 22.5 px;    
-    padding: 6px 12px;
-    text-align: center;
-    vertical-align: middle;
-    @include mini {
-        font-size: 1.25rem;
-        padding: 1.25rem;
-    }  
-}
-.Btn {
-    @extend .Disabled;
-    background-color: $green;
-    transition: 0.3s ease;      
-    -webkit-appearance: button;
 
-    &:focus {
-        box-shadow: none;
-    }
-
-    &:hover {
-        background-color: $teal;
-    }
-}
