+++ src/components/nextButton/nextButton.component.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import cx from 'classnames';
+
+import styles from './nextButton.module.scss';
+
+function NextButton({
+  isGameOver,
+  isNextButtonDisabled,
+  handleNextButtonClick,
+}) {
+  const nextButtonClasses = cx({
+    [styles.Hidden]: isGameOver,
+    [styles.Disabled]: isNextButtonDisabled,
+    [styles.Btn]: !isNextButtonDisabled,
+  });
+
+  return (
+    <button
+      className={nextButtonClasses}
+      type="button"
+      onClick={handleNextButtonClick}
+      disabled={isNextButtonDisabled}
+    >
+      Next Level
+    </button>
+  );
+}
+
+export default NextButton;
